Added ADS (aiming down sight, not advertisements)
Mostly for debugging purpose for now, but could be helpful for players too.

If you hold right click when holding a gun, the gun shoots a laser ahead. With "realistic_aim" on, the laser points straight ahead, but is not accurate to the mouse, especially when pointing right. When it's off it points directly to the mouse. We need to eventually decide which system works best and is easiest for players to understand, or otherwise adjust aiming system for a middleground.