Added Coyote Jumping
Coyote jumping is a QOL feature that is common in platformers that allows the player to jump shortly after leaving the ground. It was easy to implement and should make wall jumping much easier.
I also adjusted how wall clinging works. It should feel more like sticking to the wall now. Feel free to change it back.